Patches by cherry picks

+++ src/SolverInterface/DspOsiScip.h
@@ -122,7 +122,8 @@ public:
     }
 
     /** set node limit */
-    virtual void setNodeLimit(double num) {
+    virtual void setNodeLimit(int num)
+    {
         SCIPsetLongintParam(scip_->getScip(), "limits/nodes", num);
     }
